FIR/UAST: commonize UClass (entirely)
After commonizing UFunctionCallExpression, which forces commonzing
UMethod for constructors, UClass can be fully commonized now.
