+++ uast/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/FirKotlinConverter.kt
@@ -81,7 +81,7 @@ internal object FirKotlinConverter : BaseKotlinConverter {
 
                 is KtLightClass -> {
                     // TODO: differentiate enum entry
-                    el<UClass> { FirKotlinUClass.create(original, givenParent) }
+                    el<UClass> { KotlinUClass.create(original, givenParent) }
                 }
                 is KtClassOrObject -> {
                     convertClassOrObject(original, givenParent, requiredTypes).firstOrNull()
@@ -166,7 +166,7 @@ internal object FirKotlinConverter : BaseKotlinConverter {
             // File
           alternative { KotlinUFile(element, firKotlinUastPlugin) },
             // Facade
-          alternative { element.findFacadeClass()?.let { FirKotlinUClass.create(it, givenParent) } }
+          alternative { element.findFacadeClass()?.let { KotlinUClass.create(it, givenParent) } }
         )
     }
 
@@ -176,7 +176,7 @@ internal object FirKotlinConverter : BaseKotlinConverter {
         requiredTypes: Array<out Class<out UElement>>
     ): Sequence<UElement> {
         val ktLightClass = element.toLightClass() ?: return emptySequence()
-        val uClass = FirKotlinUClass.create(ktLightClass, givenParent)
+        val uClass = KotlinUClass.create(ktLightClass, givenParent)
         return requiredTypes.accommodate(
             // Class
             alternative { uClass },
@@ -435,7 +435,7 @@ internal object FirKotlinConverter : BaseKotlinConverter {
                 is KtClassOrObject -> expr<UDeclarationsExpression> {
                     expression.toLightClass()?.let { lightClass ->
                         KotlinUDeclarationsExpression(givenParent).apply {
-                            declarations = listOf(FirKotlinUClass.create(lightClass, this))
+                            declarations = listOf(KotlinUClass.create(lightClass, this))
                         }
                     } ?: UastEmptyExpression(givenParent)
                 }
