+++ uast/uast-kotlin-fir/testData/declaration/objects.comments.fir.txt
@@ -1,6 +1,6 @@
 KotlinUFile(
     UComment(// file comment)
-    FirKotlinUClass(
+    KotlinUClass(
         KotlinUField(
             UComment(// Single-line comment bound to top-level property)
             KotlinUAnonymousClass(
@@ -14,7 +14,7 @@ KotlinUFile(
         FirKotlinUMethod(
         )
     )
-    FirKotlinUClass(
+    KotlinUClass(
         UComment(// Single-line comment bound to object)
         KotlinUField(
             UComment(// Single-line comment bound to object)
